move parts of util.rs to input/util.rs
Just copy and paste for now, so review of this commit alone is quite nice.
