+++ crates/bindings-macro/src/input.rs
@@ -1,1 +1,2 @@
 pub mod table;
+pub mod util;
